devsecops/DAST/hook/requirements.txt
selenium
python-dotenv
beautifulsoup4
